+++ components/datetime/src/fieldset.rs
@@ -35,7 +35,6 @@ pub use crate::combo::Combo;
 
 use crate::{
     fields,
-    format::neo::*,
     options::*,
     provider::{neo::*, time_zones::tz, *},
     raw::neo::RawNeoOptions,
