+++ components/datetime/src/format/time_zone.rs
@@ -4,7 +4,7 @@
 
 //! A formatter specifically for the time zone.
 
-use super::neo::TimeZoneDataPayloadsBorrowed;
+use crate::pattern::TimeZoneDataPayloadsBorrowed;
 use crate::provider::time_zones::MetazoneId;
 use crate::{fields::FieldLength, input::ExtractedInput};
 use core::fmt;
