+++ components/datetime/src/pattern/names.rs
@@ -2,20 +2,21 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
-use super::datetime::try_write_pattern_items;
 use super::{
-    GetNameForDayPeriodError, GetNameForMonthError, GetNameForWeekdayError,
-    GetSymbolForCyclicYearError, GetSymbolForEraError, MonthPlaceholderValue,
+    DateTimePattern, DateTimePatternFormatter, GetNameForDayPeriodError, GetNameForMonthError,
+    GetNameForWeekdayError, GetSymbolForCyclicYearError, GetSymbolForEraError,
+    MonthPlaceholderValue, PatternLoadError,
 };
 use crate::external_loaders::*;
 use crate::fields::{self, Field, FieldLength, FieldSymbol};
 use crate::fieldset::dynamic::CompositeDateTimeFieldSet;
+use crate::format::datetime::try_write_pattern_items;
 use crate::input;
 use crate::input::ExtractedInput;
-use crate::neo_pattern::{DateTimePattern, DateTimePatternBorrowed};
 use crate::provider::neo::*;
 use crate::provider::pattern::PatternItem;
 use crate::provider::time_zones::tz;
+use crate::scaffold::*;
 use crate::scaffold::*;
 use crate::scaffold::{
     AllInputMarkers, DateInputMarkers, DateTimeMarkers, GetField, IsInCalendar, NeoNeverMarker,
@@ -36,228 +37,6 @@ use icu_provider::marker::NeverMarker;
 use icu_provider::prelude::*;
 use writeable::TryWriteable;
 use yoke::Yokeable;
-
-pub trait DateTimeNamesHolderTrait<M: DynamicDataMarker> {
-    type Container<Variables: PartialEq + Copy + fmt::Debug>: MaybePayload2<M, Variables>
-        + fmt::Debug;
-}
-
-impl<M: DynamicDataMarker> DateTimeNamesHolderTrait<M> for NeverMarker<()> {
-    type Container<Variables: PartialEq + Copy + fmt::Debug> = ();
-}
-
-macro_rules! impl_holder_trait {
-    ($marker:path) => {
-        impl DateTimeNamesHolderTrait<$marker> for $marker {
-            type Container<Variables: PartialEq + Copy + fmt::Debug> =
-                DateTimeNamesData2<$marker, Variables>;
-        }
-    };
-}
-
-impl_holder_trait!(YearNamesV1Marker);
-impl_holder_trait!(MonthNamesV1Marker);
-impl_holder_trait!(WeekdayNamesV1Marker);
-impl_holder_trait!(DayPeriodNamesV1Marker);
-impl_holder_trait!(tz::EssentialsV1Marker);
-impl_holder_trait!(tz::LocationsV1Marker);
-impl_holder_trait!(tz::MzGenericLongV1Marker);
-impl_holder_trait!(tz::MzGenericShortV1Marker);
-impl_holder_trait!(tz::MzSpecificLongV1Marker);
-impl_holder_trait!(tz::MzSpecificShortV1Marker);
-impl_holder_trait!(tz::MzPeriodV1Marker);
-
-#[non_exhaustive]
-pub enum MaybePayloadError2 {
-    TypeTooSpecific,
-    ConflictingField,
-}
-
-impl MaybePayloadError2 {
-    fn into_load_error(self, field: Field) -> PatternLoadError {
-        match self {
-            Self::TypeTooSpecific => PatternLoadError::TypeTooSpecific(field),
-            Self::ConflictingField => PatternLoadError::ConflictingField(field),
-        }
-    }
-}
-
-pub trait MaybePayload2<M: DynamicDataMarker, Variables> {
-    fn new_empty() -> Self;
-    fn load_put<P>(
-        &mut self,
-        provider: &P,
-        req: DataRequest,
-        variables: Variables,
-    ) -> Result<Result<(), DataError>, MaybePayloadError2>
-    where
-        P: BoundDataProvider<M> + ?Sized,
-        Self: Sized;
-    fn get(&self) -> DateTimeNamesData2Borrowed<M, Variables>;
-}
-
-pub struct DateTimeNamesData2<M: DynamicDataMarker, Variables> {
-    inner: OptionalNames<Variables, DataPayload<M>>,
-}
-
-impl<M: DynamicDataMarker, Variables> fmt::Debug for DateTimeNamesData2<M, Variables>
-where
-    Variables: fmt::Debug,
-    DataPayload<M>: fmt::Debug,
-{
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.inner.fmt(f)
-    }
-}
-
-impl<M: DynamicDataMarker, Variables> DateTimeNamesData2<M, Variables> {
-    #[inline]
-    pub(crate) fn none() -> Self {
-        Self {
-            inner: OptionalNames::None,
-        }
-    }
-}
-
-pub struct DateTimeNamesData2Borrowed<'data, M: DynamicDataMarker, Variables> {
-    inner: OptionalNames<Variables, &'data <M::DataStruct as Yokeable<'data>>::Output>,
-}
-
-impl<M: DynamicDataMarker, Variables> MaybePayload2<M, Variables>
-    for DateTimeNamesData2<M, Variables>
-where
-    Variables: PartialEq + Copy,
-{
-    #[inline]
-    fn new_empty() -> Self {
-        Self {
-            inner: OptionalNames::None,
-        }
-    }
-    fn load_put<P>(
-        &mut self,
-        provider: &P,
-        req: DataRequest,
-        variables: Variables,
-    ) -> Result<Result<(), DataError>, MaybePayloadError2>
-    where
-        P: BoundDataProvider<M> + ?Sized,
-        Self: Sized,
-    {
-        let arg_variables = variables;
-        match &self.inner {
-            OptionalNames::SingleLength { variables, .. } if arg_variables == *variables => {
-                return Ok(Ok(()));
-            }
-            OptionalNames::SingleLength { .. } => {
-                return Err(MaybePayloadError2::ConflictingField);
-            }
-            OptionalNames::None => (),
-        };
-        match provider.load_bound(req) {
-            Ok(response) => {
-                self.inner = OptionalNames::SingleLength {
-                    payload: response.payload,
-                    variables: arg_variables,
-                };
-                Ok(Ok(()))
-            }
-            Err(e) => Ok(Err(e)),
-        }
-    }
-    #[inline]
-    fn get(&self) -> DateTimeNamesData2Borrowed<M, Variables> {
-        DateTimeNamesData2Borrowed {
-            inner: self.inner.as_borrowed(),
-        }
-    }
-}
-
-impl<M: DynamicDataMarker, Variables> MaybePayload2<M, Variables> for () {
-    #[inline]
-    fn new_empty() -> Self {}
-    #[inline]
-    fn load_put<P>(
-        &mut self,
-        _: &P,
-        _: DataRequest,
-        _: Variables,
-    ) -> Result<Result<(), DataError>, MaybePayloadError2>
-    where
-        P: BoundDataProvider<M> + ?Sized,
-        Self: Sized,
-    {
-        Err(MaybePayloadError2::TypeTooSpecific)
-    }
-    #[allow(clippy::needless_lifetimes)] // Yokeable is involved
-    #[inline]
-    fn get(&self) -> DateTimeNamesData2Borrowed<M, Variables> {
-        DateTimeNamesData2Borrowed {
-            inner: OptionalNames::None,
-        }
-    }
-}
-
-/// This can be extended in the future to support multiple lengths.
-/// For now, this type wraps a symbols object tagged with a single length. See #4337
-#[derive(Debug, Copy, Clone)]
-enum OptionalNames<Variables, Payload> {
-    None,
-    SingleLength {
-        variables: Variables,
-        payload: Payload,
-    },
-}
-
-impl<Variables, Payload> OptionalNames<Variables, Payload>
-where
-    Variables: Copy + PartialEq,
-    Payload: Copy,
-{
-    pub(crate) fn get_with_variables(&self, arg_variables: Variables) -> Option<Payload> {
-        match self {
-            Self::None => None,
-            Self::SingleLength { variables, payload } if arg_variables == *variables => {
-                Some(*payload)
-            }
-            _ => None,
-        }
-    }
-}
-
-impl<Payload> OptionalNames<(), Payload>
-where
-    Payload: Copy,
-{
-    pub(crate) fn get_option(&self) -> Option<Payload> {
-        match self {
-            Self::SingleLength {
-                variables: (),
-                payload,
-            } => Some(*payload),
-            _ => None,
-        }
-    }
-}
-
-impl<M: DynamicDataMarker, Variables> OptionalNames<Variables, DataPayload<M>>
-where
-    Variables: Copy,
-{
-    #[allow(clippy::needless_lifetimes)] // Yokeable is involved
-    #[inline]
-    pub(crate) fn as_borrowed<'a>(
-        &'a self,
-    ) -> OptionalNames<Variables, &'a <M::DataStruct as Yokeable<'a>>::Output> {
-        match self {
-            Self::None => OptionalNames::None,
-            Self::SingleLength { variables, payload } => OptionalNames::SingleLength {
-                variables: *variables,
-                payload: payload.get(),
-            },
-        }
-    }
-}
 
 size_test!(
     TypedDateTimeNames<icu_calendar::Gregorian, DateTimeMarker>,
@@ -405,20 +184,6 @@ pub struct TypedDateTimeNames<C: CldrCalendar, R: DateTimeNamesMarker = Composit
     locale: DataLocale,
     inner: RawDateTimeNames<R>,
     _calendar: PhantomData<C>,
-}
-
-pub trait DateTimeNamesMarker {
-    type YearNames: DateTimeNamesHolderTrait<YearNamesV1Marker>;
-    type MonthNames: DateTimeNamesHolderTrait<MonthNamesV1Marker>;
-    type WeekdayNames: DateTimeNamesHolderTrait<WeekdayNamesV1Marker>;
-    type DayPeriodNames: DateTimeNamesHolderTrait<DayPeriodNamesV1Marker>;
-    type ZoneEssentials: DateTimeNamesHolderTrait<tz::EssentialsV1Marker>;
-    type ZoneLocations: DateTimeNamesHolderTrait<tz::LocationsV1Marker>;
-    type ZoneGenericLong: DateTimeNamesHolderTrait<tz::MzGenericLongV1Marker>;
-    type ZoneGenericShort: DateTimeNamesHolderTrait<tz::MzGenericShortV1Marker>;
-    type ZoneSpecificLong: DateTimeNamesHolderTrait<tz::MzSpecificLongV1Marker>;
-    type ZoneSpecificShort: DateTimeNamesHolderTrait<tz::MzSpecificShortV1Marker>;
-    type MetazoneLookup: DateTimeNamesHolderTrait<tz::MzPeriodV1Marker>;
 }
 
 #[derive(Debug)]
@@ -1469,11 +1234,7 @@ impl<C: CldrCalendar, R: DateTimeNamesMarker> TypedDateTimeNames<C, R> {
         &'l self,
         pattern: &'l DateTimePattern,
     ) -> DateTimePatternFormatter<'l, C, R> {
-        DateTimePatternFormatter {
-            inner: self.inner.with_pattern(pattern.as_borrowed()),
-            _calendar: PhantomData,
-            _marker: PhantomData,
-        }
+        DateTimePatternFormatter::new(pattern.as_borrowed(), self.inner.as_borrowed())
     }
 
     /// Associates this [`TypedDateTimeNames`] with a datetime pattern
@@ -1519,11 +1280,10 @@ impl<C: CldrCalendar, R: DateTimeNamesMarker> TypedDateTimeNames<C, R> {
             locale,
             pattern.iter_items(),
         )?;
-        Ok(DateTimePatternFormatter {
-            inner: self.inner.with_pattern(pattern.as_borrowed()),
-            _calendar: PhantomData,
-            _marker: PhantomData,
-        })
+        Ok(DateTimePatternFormatter::new(
+            pattern.as_borrowed(),
+            self.inner.as_borrowed(),
+        ))
     }
 
     /// Associates this [`TypedDateTimeNames`] with a pattern
@@ -1590,38 +1350,11 @@ impl<C: CldrCalendar, R: DateTimeNamesMarker> TypedDateTimeNames<C, R> {
             locale,
             pattern.iter_items(),
         )?;
-        Ok(DateTimePatternFormatter {
-            inner: self.inner.with_pattern(pattern.as_borrowed()),
-            _calendar: PhantomData,
-            _marker: PhantomData,
-        })
+        Ok(DateTimePatternFormatter::new(
+            pattern.as_borrowed(),
+            self.inner.as_borrowed(),
+        ))
     }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, displaydoc::Display)]
-/// Error returned from [`TypedDateTimeNames`]'s pattern load methods.
-#[non_exhaustive]
-pub enum PatternLoadError {
-    /// A field conflicts with a previous field.
-    ///
-    /// Fields conflict if they require the same type of data, for example the
-    /// `EEE` and `EEEE` fields (short vs long weekday) conflict, or the `M`
-    /// and `L` (format vs standalone month) conflict.
-    #[displaydoc("A field conflicts with a previous field.")]
-    ConflictingField(Field),
-    /// The field symbol is not supported in that length.
-    ///
-    /// Some fields, such as `O` are not defined for all lengths (e.g. `OO`).
-    #[displaydoc("The field symbol is not supported in that length.")]
-    UnsupportedLength(Field),
-    /// The specific type does not support this field.
-    ///
-    /// This happens for example when trying to load a month field
-    /// on a [`TypedDateTimeNames<Gregorian, ZoneFieldSet>`].
-    #[displaydoc("The specific type does not support this field.")]
-    TypeTooSpecific(Field),
-    /// An error arising from the [`DataProvider`].
-    Data(DataError),
 }
 
 impl<R: DateTimeNamesMarker> RawDateTimeNames<R> {
@@ -2001,19 +1734,6 @@ impl<R: DateTimeNamesMarker> RawDateTimeNames<R> {
         Ok(())
     }
 
-    /// Associates this [`TypedDateTimeNames`] with a pattern
-    /// without loading additional data for that pattern.
-    #[inline]
-    pub(crate) fn with_pattern<'l>(
-        &'l self,
-        pattern: DateTimePatternBorrowed<'l>,
-    ) -> RawDateTimePatternFormatter<'l> {
-        RawDateTimePatternFormatter {
-            pattern,
-            names: self.as_borrowed(),
-        }
-    }
-
     /// Loads all data required for formatting the given [`PatternItem`]s.
     ///
     /// This function has a lot of arguments because many of the arguments are generic,
@@ -2220,257 +1940,6 @@ impl<R: DateTimeNamesMarker> RawDateTimeNames<R> {
     }
 }
 
-#[derive(Debug, Copy, Clone)]
-pub struct DateTimePatternFormatter<'a, C: CldrCalendar, R> {
-    inner: RawDateTimePatternFormatter<'a>,
-    _calendar: PhantomData<C>,
-    _marker: PhantomData<R>,
-}
-
-#[derive(Debug, Copy, Clone)]
-pub(crate) struct RawDateTimePatternFormatter<'a> {
-    pattern: DateTimePatternBorrowed<'a>,
-    names: RawDateTimeNamesBorrowed<'a>,
-}
-
-impl<'a, C: CldrCalendar, R: DateTimeMarkers> DateTimePatternFormatter<'a, C, R>
-where
-    R::D: TypedDateDataMarkers<C> + DateInputMarkers,
-    R::T: TimeMarkers,
-    R::Z: ZoneMarkers,
-{
-    /// Formats a date and time of day.
-    ///
-    /// For an example, see [`TypedDateTimeNames`].
-    pub fn format<I>(&self, datetime: &I) -> FormattedDateTimePattern<'a>
-    where
-        I: ?Sized + IsInCalendar<C> + AllInputMarkers<R>,
-    {
-        FormattedDateTimePattern {
-            pattern: self.inner.pattern,
-            input: ExtractedInput::extract_from_neo_input::<R::D, R::T, R::Z, I>(datetime),
-            names: self.inner.names,
-        }
-    }
-
-    /// Formats a date without a time of day.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use icu::calendar::Date;
-    /// use icu::calendar::Gregorian;
-    /// use icu::datetime::fields;
-    /// use icu::datetime::fields::FieldLength;
-    /// use icu::datetime::neo_pattern::DateTimePattern;
-    /// use icu::datetime::TypedDateTimeNames;
-    /// use icu::locale::locale;
-    /// use writeable::assert_try_writeable_eq;
-    ///
-    /// // Create an instance that can format wide month and era names:
-    /// let mut names: TypedDateTimeNames<Gregorian> =
-    ///     TypedDateTimeNames::try_new(&locale!("en-GB").into()).unwrap();
-    /// names
-    ///     .include_month_names(fields::Month::Format, FieldLength::Four)
-    ///     .unwrap()
-    ///     .include_year_names(FieldLength::Four)
-    ///     .unwrap();
-    ///
-    /// // Create a pattern from a pattern string:
-    /// let pattern_str = "'The date is:' MMMM d, y GGGG";
-    /// let pattern: DateTimePattern = pattern_str.parse().unwrap();
-    ///
-    /// // Test it with some different dates:
-    /// // Note: extended year -50 is year 51 BCE
-    /// let date_bce = Date::try_new_gregorian(-50, 3, 15).unwrap();
-    /// let date_ce = Date::try_new_gregorian(1700, 11, 20).unwrap();
-    /// assert_try_writeable_eq!(
-    ///     names.with_pattern(&pattern).format_date(&date_bce),
-    ///     "The date is: March 15, 51 Before Christ"
-    /// );
-    /// assert_try_writeable_eq!(
-    ///     names.with_pattern(&pattern).format_date(&date_ce),
-    ///     "The date is: November 20, 1700 Anno Domini"
-    /// );
-    /// ```
-    pub fn format_date<I>(&self, datetime: &'a I) -> FormattedDateTimePattern<'a>
-    where
-        I: ?Sized
-            + IsInCalendar<C>
-            + GetField<<R::D as DateInputMarkers>::YearInput>
-            + GetField<<R::D as DateInputMarkers>::MonthInput>
-            + GetField<<R::D as DateInputMarkers>::DayOfMonthInput>
-            + GetField<<R::D as DateInputMarkers>::DayOfWeekInput>
-            + GetField<<R::D as DateInputMarkers>::DayOfYearInput>
-            + GetField<()>,
-    {
-        FormattedDateTimePattern {
-            pattern: self.inner.pattern,
-            input: ExtractedInput::extract_from_neo_input::<R::D, NeoNeverMarker, NeoNeverMarker, I>(
-                datetime,
-            ),
-            names: self.inner.names,
-        }
-    }
-
-    /// Formats a time of day without a date.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use icu::calendar::Gregorian;
-    /// use icu::calendar::Time;
-    /// use icu::datetime::fields::FieldLength;
-    /// use icu::datetime::neo_pattern::DateTimePattern;
-    /// use icu::datetime::TypedDateTimeNames;
-    /// use icu::locale::locale;
-    /// use writeable::assert_try_writeable_eq;
-    ///
-    /// // Create an instance that can format abbreviated day periods:
-    /// let mut names: TypedDateTimeNames<Gregorian> =
-    ///     TypedDateTimeNames::try_new(&locale!("en-US").into()).unwrap();
-    /// names
-    ///     .include_day_period_names(FieldLength::Three)
-    ///     .unwrap();
-    ///
-    /// // Create a pattern from a pattern string:
-    /// let pattern_str = "'The time is:' h:mm b";
-    /// let pattern: DateTimePattern = pattern_str.parse().unwrap();
-    ///
-    /// // Test it with different times of day:
-    /// let time_am = Time::try_new(11, 4, 14, 0).unwrap();
-    /// let time_pm = Time::try_new(13, 41, 28, 0).unwrap();
-    /// let time_noon = Time::try_new(12, 0, 0, 0).unwrap();
-    /// let time_midnight = Time::try_new(0, 0, 0, 0).unwrap();
-    /// assert_try_writeable_eq!(
-    ///     names.with_pattern(&pattern).format_time(&time_am),
-    ///     "The time is: 11:04 AM"
-    /// );
-    /// assert_try_writeable_eq!(
-    ///     names.with_pattern(&pattern).format_time(&time_pm),
-    ///     "The time is: 1:41 PM"
-    /// );
-    /// assert_try_writeable_eq!(
-    ///     names.with_pattern(&pattern).format_time(&time_noon),
-    ///     "The time is: 12:00 noon"
-    /// );
-    /// assert_try_writeable_eq!(
-    ///     names.with_pattern(&pattern).format_time(&time_midnight),
-    ///     "The time is: 12:00 midnight"
-    /// );
-    /// ```
-    pub fn format_time<I>(&self, datetime: &'a I) -> FormattedDateTimePattern<'a>
-    where
-        I: ?Sized
-            + IsInCalendar<C>
-            + GetField<<R::T as TimeMarkers>::HourInput>
-            + GetField<<R::T as TimeMarkers>::MinuteInput>
-            + GetField<<R::T as TimeMarkers>::SecondInput>
-            + GetField<<R::T as TimeMarkers>::NanoSecondInput>
-            + GetField<()>,
-    {
-        FormattedDateTimePattern {
-            pattern: self.inner.pattern,
-            input: ExtractedInput::extract_from_neo_input::<NeoNeverMarker, R::T, NeoNeverMarker, I>(
-                datetime,
-            ),
-            names: self.inner.names,
-        }
-    }
-
-    /// Formats a timezone without a date or time.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use icu::calendar::Gregorian;
-    /// use icu::datetime::neo_pattern::DateTimePattern;
-    /// use icu::datetime::fieldset::dynamic::ZoneFieldSet;
-    /// use icu::datetime::TypedDateTimeNames;
-    /// use icu::locale::locale;
-    /// use icu::timezone::IxdtfParser;
-    /// use writeable::assert_try_writeable_eq;
-    ///
-    /// let mut london_winter = IxdtfParser::new().try_from_str(
-    ///     "2024-01-01T00:00:00+00:00[Europe/London]",
-    /// )
-    /// .unwrap()
-    /// .to_calendar(Gregorian);
-    /// let mut london_summer = IxdtfParser::new().try_from_str(
-    ///     "2024-07-01T00:00:00+01:00[Europe/London]",
-    /// )
-    /// .unwrap()
-    /// .to_calendar(Gregorian);
-    ///
-    /// let mut names =
-    ///     TypedDateTimeNames::<Gregorian, ZoneFieldSet>::try_new(
-    ///         &locale!("en-GB").into(),
-    ///     )
-    ///     .unwrap();
-    ///
-    /// names.include_time_zone_essentials().unwrap();
-    /// names.include_time_zone_specific_short_names().unwrap();
-    ///
-    /// // Create a pattern with symbol `z`:
-    /// let pattern_str = "'Your time zone is:' z";
-    /// let pattern: DateTimePattern = pattern_str.parse().unwrap();
-    ///
-    /// assert_try_writeable_eq!(
-    ///     names.with_pattern(&pattern).format_timezone(&london_winter),
-    ///     "Your time zone is: GMT",
-    /// );
-    /// assert_try_writeable_eq!(
-    ///     names.with_pattern(&pattern).format_timezone(&london_summer),
-    ///     "Your time zone is: BST",
-    /// );
-    /// ```
-    pub fn format_timezone<I>(&self, datetime: &'a I) -> FormattedDateTimePattern<'a>
-    where
-        I: ?Sized
-            + IsInCalendar<C>
-            + GetField<<R::Z as ZoneMarkers>::TimeZoneIdInput>
-            + GetField<<R::Z as ZoneMarkers>::TimeZoneOffsetInput>
-            + GetField<<R::Z as ZoneMarkers>::TimeZoneVariantInput>
-            + GetField<<R::Z as ZoneMarkers>::TimeZoneLocalTimeInput>
-            + GetField<()>,
-    {
-        FormattedDateTimePattern {
-            pattern: self.inner.pattern,
-            input: ExtractedInput::extract_from_neo_input::<NeoNeverMarker, NeoNeverMarker, R::Z, I>(
-                datetime,
-            ),
-            names: self.inner.names,
-        }
-    }
-}
-
-/// A pattern that has been interpolated and implements [`TryWriteable`].
-#[derive(Debug)]
-pub struct FormattedDateTimePattern<'a> {
-    pattern: DateTimePatternBorrowed<'a>,
-    input: ExtractedInput,
-    names: RawDateTimeNamesBorrowed<'a>,
-}
-
-impl TryWriteable for FormattedDateTimePattern<'_> {
-    type Error = DateTimeWriteError;
-    fn try_write_to_parts<S: writeable::PartsWrite + ?Sized>(
-        &self,
-        sink: &mut S,
-    ) -> Result<Result<(), Self::Error>, fmt::Error> {
-        try_write_pattern_items(
-            self.pattern.0.as_borrowed().metadata,
-            self.pattern.0.as_borrowed().items.iter(),
-            &self.input,
-            &self.names,
-            self.names.fixed_decimal_formatter,
-            sink,
-        )
-    }
-
-    // TODO(#489): Implement writeable_length_hint
-}
-
 impl<'data> RawDateTimeNamesBorrowed<'data> {
     pub(crate) fn get_name_for_month(
         &self,
@@ -2655,376 +2124,6 @@ impl<'data> RawDateTimeNamesBorrowed<'data> {
             mz_specific_long: self.mz_specific_long.get_option(),
             mz_specific_short: self.mz_specific_short.get_option(),
             mz_periods: self.mz_periods.get_option(),
-        }
-    }
-}
-
-#[cfg(test)]
-#[cfg(feature = "compiled_data")]
-mod tests {
-    use super::*;
-    use icu_calendar::{DateTime, Gregorian};
-    use icu_locale_core::locale;
-    use writeable::assert_try_writeable_eq;
-
-    #[test]
-    fn test_basic_pattern_formatting() {
-        let locale = locale!("en").into();
-        let mut names: TypedDateTimeNames<Gregorian> =
-            TypedDateTimeNames::try_new(&locale).unwrap();
-        names
-            .load_month_names(
-                &crate::provider::Baked,
-                fields::Month::Format,
-                fields::FieldLength::Four,
-            )
-            .unwrap()
-            .load_weekday_names(
-                &crate::provider::Baked,
-                fields::Weekday::Format,
-                fields::FieldLength::Three,
-            )
-            .unwrap()
-            .load_year_names(&crate::provider::Baked, FieldLength::Five)
-            .unwrap()
-            .load_day_period_names(&crate::provider::Baked, FieldLength::Three)
-            .unwrap();
-        let pattern: DateTimePattern = "'It is' E, MMMM d, y GGGGG 'at' hh:mm a'!'"
-            .parse()
-            .unwrap();
-        let datetime = DateTime::try_new_gregorian(2023, 10, 25, 15, 0, 55).unwrap();
-        let formatted_pattern = names.with_pattern(&pattern).format(&datetime);
-
-        assert_try_writeable_eq!(
-            formatted_pattern,
-            "It is Wed, October 25, 2023 A at 03:00 PM!",
-            Ok(()),
-        );
-    }
-
-    #[test]
-    fn test_era_coverage() {
-        let locale = locale!("uk").into();
-        #[derive(Debug)]
-        struct TestCase {
-            pattern: &'static str,
-            field_length: FieldLength,
-            expected: &'static str,
-        }
-        let cases = [
-            TestCase {
-                pattern: "<G>",
-                field_length: FieldLength::Three,
-                expected: "<н. е.>",
-            },
-            TestCase {
-                pattern: "<GG>",
-                field_length: FieldLength::Three,
-                expected: "<н. е.>",
-            },
-            TestCase {
-                pattern: "<GGG>",
-                field_length: FieldLength::Three,
-                expected: "<н. е.>",
-            },
-            TestCase {
-                pattern: "<GGGG>",
-                field_length: FieldLength::Four,
-                expected: "<нашої ери>",
-            },
-            TestCase {
-                pattern: "<GGGGG>",
-                field_length: FieldLength::Five,
-                expected: "<н.е.>",
-            },
-        ];
-        for cas in cases {
-            let TestCase {
-                pattern,
-                field_length,
-                expected,
-            } = cas;
-            let mut names: TypedDateTimeNames<Gregorian> =
-                TypedDateTimeNames::try_new(&locale).unwrap();
-            names
-                .load_year_names(&crate::provider::Baked, field_length)
-                .unwrap();
-            let pattern: DateTimePattern = pattern.parse().unwrap();
-            let datetime = DateTime::try_new_gregorian(2023, 11, 17, 13, 41, 28).unwrap();
-            let formatted_pattern = names.with_pattern(&pattern).format(&datetime);
-
-            assert_try_writeable_eq!(formatted_pattern, expected, Ok(()), "{cas:?}");
-        }
-    }
-
-    #[test]
-    fn test_month_coverage() {
-        // Ukrainian has different values for format and standalone
-        let locale = locale!("uk").into();
-        #[derive(Debug)]
-        struct TestCase {
-            pattern: &'static str,
-            field_symbol: fields::Month,
-            field_length: FieldLength,
-            expected: &'static str,
-        }
-        let cases = [
-            // 'M' and 'MM' are numeric
-            TestCase {
-                pattern: "<MMM>",
-                field_symbol: fields::Month::Format,
-                field_length: FieldLength::Three,
-                expected: "<лист.>",
-            },
-            TestCase {
-                pattern: "<MMMM>",
-                field_symbol: fields::Month::Format,
-                field_length: FieldLength::Four,
-                expected: "<листопада>",
-            },
-            TestCase {
-                pattern: "<MMMMM>",
-                field_symbol: fields::Month::Format,
-                field_length: FieldLength::Five,
-                expected: "<л>",
-            },
-            // 'L' and 'LL' are numeric
-            TestCase {
-                pattern: "<LLL>",
-                field_symbol: fields::Month::StandAlone,
-                field_length: FieldLength::Three,
-                expected: "<лист.>",
-            },
-            TestCase {
-                pattern: "<LLLL>",
-                field_symbol: fields::Month::StandAlone,
-                field_length: FieldLength::Four,
-                expected: "<листопад>",
-            },
-            TestCase {
-                pattern: "<LLLLL>",
-                field_symbol: fields::Month::StandAlone,
-                field_length: FieldLength::Five,
-                expected: "<Л>",
-            },
-        ];
-        for cas in cases {
-            let TestCase {
-                pattern,
-                field_symbol,
-                field_length,
-                expected,
-            } = cas;
-            let mut names: TypedDateTimeNames<Gregorian> =
-                TypedDateTimeNames::try_new(&locale).unwrap();
-            names
-                .load_month_names(&crate::provider::Baked, field_symbol, field_length)
-                .unwrap();
-            let pattern: DateTimePattern = pattern.parse().unwrap();
-            let datetime = DateTime::try_new_gregorian(2023, 11, 17, 13, 41, 28).unwrap();
-            let formatted_pattern = names.with_pattern(&pattern).format(&datetime);
-
-            assert_try_writeable_eq!(formatted_pattern, expected, Ok(()), "{cas:?}");
-        }
-    }
-
-    #[test]
-    fn test_weekday_coverage() {
-        let locale = locale!("uk").into();
-        #[derive(Debug)]
-        struct TestCase {
-            pattern: &'static str,
-            field_symbol: fields::Weekday,
-            field_length: FieldLength,
-            expected: &'static str,
-        }
-        let cases = [
-            TestCase {
-                pattern: "<E>",
-                field_symbol: fields::Weekday::Format,
-                field_length: FieldLength::Three,
-                expected: "<пт>",
-            },
-            TestCase {
-                pattern: "<EE>",
-                field_symbol: fields::Weekday::Format,
-                field_length: FieldLength::Three,
-                expected: "<пт>",
-            },
-            TestCase {
-                pattern: "<EEE>",
-                field_symbol: fields::Weekday::Format,
-                field_length: FieldLength::Three,
-                expected: "<пт>",
-            },
-            TestCase {
-                pattern: "<EEEE>",
-                field_symbol: fields::Weekday::Format,
-                field_length: FieldLength::Four,
-                expected: "<пʼятниця>",
-            },
-            TestCase {
-                pattern: "<EEEEE>",
-                field_symbol: fields::Weekday::Format,
-                field_length: FieldLength::Five,
-                expected: "<П>",
-            },
-            TestCase {
-                pattern: "<EEEEEE>",
-                field_symbol: fields::Weekday::Format,
-                field_length: FieldLength::Six,
-                expected: "<пт>",
-            },
-            // 'e' and 'ee' are numeric
-            TestCase {
-                pattern: "<eee>",
-                field_symbol: fields::Weekday::Format,
-                field_length: FieldLength::Three,
-                expected: "<пт>",
-            },
-            TestCase {
-                pattern: "<eeee>",
-                field_symbol: fields::Weekday::Format,
-                field_length: FieldLength::Four,
-                expected: "<пʼятниця>",
-            },
-            TestCase {
-                pattern: "<eeeee>",
-                field_symbol: fields::Weekday::Format,
-                field_length: FieldLength::Five,
-                expected: "<П>",
-            },
-            TestCase {
-                pattern: "<eeeeee>",
-                field_symbol: fields::Weekday::Format,
-                field_length: FieldLength::Six,
-                expected: "<пт>",
-            },
-            // 'c' and 'cc' are numeric
-            TestCase {
-                pattern: "<ccc>",
-                field_symbol: fields::Weekday::StandAlone,
-                field_length: FieldLength::Three,
-                expected: "<пт>",
-            },
-            TestCase {
-                pattern: "<cccc>",
-                field_symbol: fields::Weekday::StandAlone,
-                field_length: FieldLength::Four,
-                expected: "<пʼятниця>",
-            },
-            TestCase {
-                pattern: "<ccccc>",
-                field_symbol: fields::Weekday::StandAlone,
-                field_length: FieldLength::Five,
-                expected: "<П>",
-            },
-            TestCase {
-                pattern: "<cccccc>",
-                field_symbol: fields::Weekday::StandAlone,
-                field_length: FieldLength::Six,
-                expected: "<пт>",
-            },
-        ];
-        for cas in cases {
-            let TestCase {
-                pattern,
-                field_symbol,
-                field_length,
-                expected,
-            } = cas;
-            let mut names: TypedDateTimeNames<Gregorian> =
-                TypedDateTimeNames::try_new(&locale).unwrap();
-            names
-                .load_weekday_names(&crate::provider::Baked, field_symbol, field_length)
-                .unwrap();
-            let pattern: DateTimePattern = pattern.parse().unwrap();
-            let datetime = DateTime::try_new_gregorian(2023, 11, 17, 13, 41, 28).unwrap();
-            let formatted_pattern = names.with_pattern(&pattern).format(&datetime);
-
-            assert_try_writeable_eq!(formatted_pattern, expected, Ok(()), "{cas:?}");
-        }
-    }
-
-    #[test]
-    fn test_dayperiod_coverage() {
-        // Thai has different values for different lengths of day periods
-        // TODO(#487): Support flexible day periods, too
-        let locale = locale!("th").into();
-        #[derive(Debug)]
-        struct TestCase {
-            pattern: &'static str,
-            field_length: FieldLength,
-            expected: &'static str,
-        }
-        let cases = [
-            TestCase {
-                pattern: "<a>",
-                field_length: FieldLength::Three,
-                expected: "<PM>",
-            },
-            TestCase {
-                pattern: "<aa>",
-                field_length: FieldLength::Three,
-                expected: "<PM>",
-            },
-            TestCase {
-                pattern: "<aaa>",
-                field_length: FieldLength::Three,
-                expected: "<PM>",
-            },
-            TestCase {
-                pattern: "<aaaa>",
-                field_length: FieldLength::Four,
-                expected: "<หลังเที่ยง>",
-            },
-            TestCase {
-                pattern: "<aaaaa>",
-                field_length: FieldLength::Five,
-                expected: "<p>",
-            },
-            TestCase {
-                pattern: "<b>",
-                field_length: FieldLength::Three,
-                expected: "<PM>",
-            },
-            TestCase {
-                pattern: "<bb>",
-                field_length: FieldLength::Three,
-                expected: "<PM>",
-            },
-            TestCase {
-                pattern: "<bbb>",
-                field_length: FieldLength::Three,
-                expected: "<PM>",
-            },
-            TestCase {
-                pattern: "<bbbb>",
-                field_length: FieldLength::Four,
-                expected: "<หลังเที่ยง>",
-            },
-            TestCase {
-                pattern: "<bbbbb>",
-                field_length: FieldLength::Five,
-                expected: "<p>",
-            },
-        ];
-        for cas in cases {
-            let TestCase {
-                pattern,
-                field_length,
-                expected,
-            } = cas;
-            let mut names: TypedDateTimeNames<Gregorian> =
-                TypedDateTimeNames::try_new(&locale).unwrap();
-            names
-                .load_day_period_names(&crate::provider::Baked, field_length)
-                .unwrap();
-            let pattern: DateTimePattern = pattern.parse().unwrap();
-            let datetime = DateTime::try_new_gregorian(2023, 11, 17, 13, 41, 28).unwrap();
-            let formatted_pattern = names.with_pattern(&pattern).format(&datetime);
-
-            assert_try_writeable_eq!(formatted_pattern, expected, Ok(()), "{cas:?}");
         }
     }
 }
