+++ components/datetime/src/scaffold/fieldset_traits.rs
@@ -2,8 +2,9 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
+use core::fmt;
+
 use crate::{
-    format::neo::*,
     provider::{neo::*, time_zones::tz, *},
     scaffold::*,
 };
